python/beerbar/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-factory-boy==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==24.8.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
